all: switch to dual-licensing, GPLv2 or GPLv3
I'm the only author and Copyright holder for all those scripts, so I can
relicense them as I will.

As a (arguably weak) proof, the history of those scripts is stored in
the file HISTORY.git-log, retrieved from my home repository, dating back
to 2012-11-27T18:29:57+0000. This will allow to compare such a log in
the future, thanks to the sha1 for each commit.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>