git-pwv: properly detect end-of-body
We currently detect the end of the body whith the --- line.

However, a submitter may add his own --- line to indicate end-of-commit
log, but provide personal messages like revision history of the patch,
or some random comments not to be part of the commit log.

Detect the end-of-body whith the start of the diffstat, not when we
encounter the --- line.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>