git-send-pull-request: workaround deffective git-pull-request
In the past, git-pull-request was very happy with branches and tags that
had a slash '/' in their neames.

That's no longer the case with recent-ish git version, at least the one
packaged in Ubuntu 16.04.

Fix that by replacing git-pull-request with a hand-crafted equivalent
that should look (almost) exactly the same.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>