git-send-pull-request: add --force to always send
When the remote is fetched from a caching server (e.g. an http server
that caches the git stuff, boo!), it might appear to us that it is not
up-to-date, even if a push was just done, until the server refreshes its
cache. Which can take quite a biut of time.

In this case, git-send-pull-request currently refuses to send the pull
request, on the assumption that the remote is not up-to-date.

Add a --force option to turn this error into a mere warning.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>